+++ common/sampling.cpp
@@ -2,6 +2,7 @@
 
 #include "common.h"
 #include "log.h"
+#include "reasoning-budget.h"
 
 #include <algorithm>
 #include <cmath>
@@ -252,11 +253,11 @@ struct common_sampler * common_sampler_init(const struct llama_model * model, st
 
     // reasoning budget sampler — added first so it can force tokens before other samplers
     if (params.reasoning_budget_tokens >= 0 && !params.reasoning_budget_forced.empty()) {
-        samplers.push_back(llama_sampler_init_reasoning_budget(
+        samplers.push_back(common_reasoning_budget_init(
             vocab,
-            params.reasoning_budget_start.data(),  params.reasoning_budget_start.size(),
-            params.reasoning_budget_end.data(),    params.reasoning_budget_end.size(),
-            params.reasoning_budget_forced.data(), params.reasoning_budget_forced.size(),
+            params.reasoning_budget_start,
+            params.reasoning_budget_end,
+            params.reasoning_budget_forced,
             params.reasoning_budget_tokens,
             params.reasoning_budget_activate_immediately));
     }
